# docs/openapi.yaml
openapi: 3.0.0
info:
  version: "1.0.0"
  title: MoneyMinder - Rest api
  description: "Api for MoneyMinder project"
...
paths:
...
  /categories/{categoryId}:
    put:
      security:
        - bearerAuth: []
      tags:
        - categories
      description: Update category
      operationId: updateCategory
      requestBody:
        description: Update category body
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryRequestBody'
      parameters:
        - in: path
          name: categoryId
          schema:
            type: integer
          required: true
          description: ID of the category
      responses:
        '200':
          description: Category updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          description: Category not found
        '404':
          description: Wrong request body

    delete:
      security:
        - bearerAuth: []
      tags:
        - categories
      description: Delete specific category
      operationId: deleteCategory
      parameters:
        - in: path
          name: categoryId
          schema:
            type: integer
          required: true
          description: ID of the category
      responses:
        '200':
          description: Category deleted successfully
        '404':
          description: Category not found


    get:
      security:
        - bearerAuth: []
      tags:
        - categories
      description: Get specific category
      operationId: getSpecificCategory
      parameters:
        - in: path
          name: categoryId
          schema:
            type: integer
          required: true
          description: ID of the category
      responses:
        '200':
          description: Returned specific category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          description: Category not found
        '404':
          description: Wrong request body
...
  /lists/{listId}:
    get:
      security:
        - bearerAuth: []
      tags:
        - lists
      description: Return specific list
      operationId: getSpecificList
      parameters:
        - in: path
          name: listId
          schema:
            type: integer
          required: true
          description: ID of the list
      responses:
        '200':
          description: Successfully returned specific list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/List'
        '404':
          description: List not found


    delete:
      security:
        - bearerAuth: []
      tags:
        - lists
      description: Delete specific list
      operationId: deleteSpecificList
      parameters:
        - in: path
          name: listId
          schema:
            type: integer
          required: true
          description: ID of the list
      responses:
        '200':
          description: Successfully deleted specific list
        '404':
          description: List not found


    put:
      security:
        - bearerAuth: []
      tags:
        - lists
      description: Update specific list
      operationId: updateSpecificList
      requestBody:
        description: Update list body
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListRequestBody'
      parameters:
        - in: path
          name: listId
          schema:
            type: integer
          required: true
          description: ID of the list
      responses:
        '200':
          description: Successfully updated specific list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/List'
        '400':
          description: Wrong request body
        '404':
          description: List not found
...
  /lists/{listId}/items/{itemId}:
    get:
      security:
        - bearerAuth: []
      tags:
        - items
      description: Return specific item from specific list
      operationId: returnSpecificItemFromSpecificList
      parameters:
        - in: path
          name: listId
          schema:
            type: integer
          required: true
          description: ID of the list
        - in: path
          name: itemId
          schema:
            type: integer
          required: true
          description: ID of the item
      responses:
        '200':
          description: Successfully returned item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '404':
          description: List or Item not found


    put:
      security:
        - bearerAuth: []
      tags:
        - items
      description: Update specific item from specific list
      operationId: UpdateSpecificItemFromSpecificList
      requestBody:
        description: Update item body
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemRequestBody'
      parameters:
        - in: path
          name: listId
          schema:
            type: integer
          required: true
          description: ID of the list
        - in: path
          name: itemId
          schema:
            type: integer
          required: true
          description: ID of the item
      responses:
        '200':
          description: Successfully updated item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '400':
          description: Wrong request body
        '404':
          description: List or Item not found


    delete:
      security:
        - bearerAuth: []
      tags:
        - items
      description: Delete specific item from specific list
      operationId: deleteSpecificItemFromSpecificList
      parameters:
        - in: path
          name: listId
          schema:
            type: integer
          required: true
          description: ID of the list
        - in: path
          name: itemId
          schema:
            type: integer
          required: true
          description: ID of the item
      responses:
        '200':
          description: Successfully deleted item
        '404':
          description: List or Item not found
...
  /purchased-items/categories/{categoryId}/days/{days}:

    get:
      security:
        - bearerAuth: []
      tags:
        - purchasedItems
      description: Get purchased items in last N days from category
      operationId: getPurchasedItemsInNDaysByCategory
      parameters:
        - in: path
          name: categoryId
          schema:
            type: integer
          required: true
          description: ID of the category
        - in: path
          name: days
          schema:
            type: integer
          required: true
          description: amount of days
      responses:
        '200':
          description: Successfully returned purchased items from n Days
        '404':
          description: Category not found
        '400':
          description: Wrong amount of days
  /purchased-items/categories/{categoryId}:

    get:
      security:
        - bearerAuth: []
      tags:
        - purchasedItems
      description: Get purchased items by categoryId
      operationId: getPurchasedItemsByCategoryId
      parameters:
        - in: path
          name: categoryId
          schema:
            type: integer
          required: true
          description: ID of the category
      responses:
        '200':
          description: Successfully returned purchased items
        '404':
          description: Category not found
  /purchased-items/names/{prefix}:

    get:
      security:
        - bearerAuth: []
      tags:
        - purchasedItems
      description: Get purchased items names by prefix
      operationId: getPurchasedItemsNamesByPrefix
      parameters:
        - in: path
          name: prefix
          schema:
            type: string
          required: true
          description: prefix
      responses:
        '200':
          description: Successfully returned names
  /purchased-items/days/{days}:

    get:
      security:
        - bearerAuth: []
      tags:
        - purchasedItems
      description: Get purchased items in N Days
      operationId: getPurchasedItemsInNDays
      parameters:
        - in: path
          name: days
          schema:
            type: integer
          required: true
          description: amount od days
      responses:
        '200':
          description: Successfully returned names
        '400':
          description: Wrong amount of days
  /purchased-items/items/{amountOfItems}:

    get:
      security:
        - bearerAuth: []
      tags:
        - purchasedItems
      description: Get last N purchased items
      operationId: getLastNPurchasedItems
      parameters:
        - in: path
          name: amountOfItems
          schema:
            type: integer
          required: true
          description: amount of items
      responses:
        '200':
          description: Successfully returned items
        '400':
          description: Wrong amount of items
  /purchased-items/lists/{listId}:

    get:
      security:
        - bearerAuth: []
      tags:
        - purchasedItems
      description: Return purchased items from list
      operationId: getLastPurchasedItemsFromList
      parameters:
        - in: path
          name: listId
          schema:
            type: integer
          required: true
          description: List id
      responses:
        '200':
          description: Successfully returned items
        '400':
          description: Wrong list id
...
components:
  schemas:
    Category:
      required:
        - name
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: 'Food'
    List:
      required:
        - name
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: 'Weekly'
    Item:
      required:
        - name
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 2
        categoryId:
          type: integer
          format: int64
          example: 3
        listId:
          type: integer
          format: int64
          example: 4
        name:
          type: string
          example: 'Cola'
        amount:
          type: integer
          format: int64
          example: 3
        price:
          type: number
          format: double
          example: 7.45
        weight:
          type: integer
          format: int64
          example: 150
        timeCreated:
          type: string
          format: date
          example: "2023-11-27"
    CategoryRequestBody:
      required:
        - name
      type: object
      properties:
        name:
          type: string
          example: 'Sweets'
    ListRequestBody:
      required:
        - name
      type: object
      properties:
        name:
          type: string
          example: 'Daily'
    ItemRequestBody:
      required:
        - name
      type: object
      properties:
        categoryId:
          type: integer
          format: int64
          example: 2
        listId:
          type: integer
          format: int64
          example: 3
        name:
          type: string
          example: 'Cola'
        amount:
          type: integer
          format: int64
          example: 5
        price:
          type: number
          format: double
          example: 4.30
        weight:
          type: integer
          format: int64
          example: 350
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT